Function Documentation

Main.Kelp.defaults — Method

Kelp.defaults(t_i, t_e, u)

Generates "default" or anayltical values for the water speed, temperature, irradiance and external nitrogen for testing.

Parameters:

t_i: start time

t_e: end time

u: your chosen water speed

Main.Kelp.equations! — Method

Kelp.equations!(y, params, t)

Equations outlined in the Main Equations section of the paper to be solved by the ODE library.

y: the current state of the system as a vector (area, nitrate reserve, carbon reserve).

params: the variable parameters of the model:

u_arr: interpolation object of the water speed in time

temp_arr: interpolation object of the temperature in time

irr_arr: interpolation object of the irradianec in time

ex_n_arr: interpolation object of the external nitrate concentration in time

NormDeltaL: the normalised change in day length

resp_model: the choice of respiration model, 1 is the origional from the 2012 paper and 2 is the modified version in 2013

dt: the time step length, this is important as it is used in the "extreme carbon limit" part of the equations, see NB.

t: the current time (with respect to the time in the interpolations)

Note: These equations must be solved with an algorithm with fixed time steps and known, constant sub timestep lengths. This is because the "extreme carbon limit" element is only implientable with known and fixed timesteps as the value of a must be changed to a particular value rather than changing the deriviative. This can only be done (within the framework of the ODE library) by setting the derivitive to (X(next)-X(old))/dt

Main.Kelp.solvekelp — Function

Kelp.solvekelp(t_i, nd, u, temp, irr, ex_n, lat, a_0, n_0, c_0, params="src/parameters/origional.jl", resp_model=1, dt=1)

Solves the model for some set of parametetrs and returns the ODE library solution as well as a dataframe of the useful results.

Parameters:

t_i: th estart time (in days since the start of the interpolation objects "day zero")

nd: the number of days to run for

u: interpolation object (IO) of water speed

temp: IO of temperature

ex_n: IO of external nitrate concentration

lat: latitude, relivant for the change of day length

a_0: initial area

n_0: initial nitrogen reserve (gN/gSW)

c_0: initial carbon reserve (gC/gSW)

params: string of the path to a parameters file, defaults to the 2012 values. Also supplied is 2013 in src/parameters/2013.jl or you can copy and vary them

resp_model: choice of respiration model, 1 (default) uses the 2012 version and 2 uses the modifcations from the 2013 paper

dt: the time step size to use (see equations! note), default is 1 day (seems small enough)

Returns:

solution: the ODE library solution

results: dataframe of area/nitrogen reserve/carbon reserve/total nitrate update. All others useful quantities can be easily derived.