Function Documentation

Main.SugarKelp.equations — Method

SugarKelp.equations(t, a, n, c, u, temp, irr, ex_n, λ, resp_model, dt)

Solves the papers main equations.

Parameters:

a: area /dm^2

n: nitrate reserve /gN/gSW

c: carbon reserve /gC/gSW

u: water speed /m/s

temp: temperature /°C

irr: irradiance / mol photons/m²/day

ex_n: external nitrate concentration /mmol m^3

λ: normalised change in day length

resp_model: choice of resparation model (see SugarKelp.r)

dt: timestep length /days

Returns:

da: results of equation 1, the area rate

dn: results of equation 7, the nitrogen rate

dc: results of equation 9, the carbon rate

j: the specific nitrate uptake rate

Main.SugarKelp.eval_j — Method

SugarKelp.eval_j(ex_n,n,u)

Solves Equation 8, the specific nitrate uptake rate.

Parameters:

ex_n: external nitrate concentration /mmol/m^3

n: nitrogen reserve /gN/gSW

u: water speed /m/s

Returns: specific nitrate uptake rate

Notes:

Names eval_j because it looks better in SugarKelp.equations to have j and you can't reuse it

Main.SugarKelp.eval_μ — Method

SugarKelp.eval_μ(a,n,c,temp,λ)

Solves Equation 2, the specific growth rate.

Parameters:

a: area /dm^2

n: nitrate reserve /gN/gSW

c: carbon reserve /gC/gSW

temp: temperature /°C

λ: normalised change in day length

Returns: specific growth rate

Notes:

Names eval_μ because it looks better in SugarKelp.equations to have μ and you can't reuse it

Main.SugarKelp.extract_valid — Method

SugarKelp.extract_valid(raw,raw_time,fill)

Extracts the valid values from an array by checking against a fill value and returns the valids and corresponding time.

Parameters:

raw: the array to check

raw_time: the corresponding time array

fill: the fill value to check against

Returns:

vals: the filtered values

time: corresponding times

Main.SugarKelp.f_area — Method

SugarKelp.f_area(a)

Solves Equation 3, the effect of area on growth

Parameters: a, area /dm^2

Returns: effect of area on growth

Main.SugarKelp.f_photo — Method

SugarKelp.f_photo(λ)

Solves Equation 5, the seasonal influence on growth rate

Parameters: λ: normalised change in day length

Returns: seasonal influence on growth

Main.SugarKelp.f_temp — Method

SugarKelp.f_temp(temp)

Solves Equation 4, the effect of temperature on growth

Parameters: temp: temperature /°C

Returns: effect of temperature on growth

Main.SugarKelp.gen_λ — Method

SugarKelp.gen_λ(lat)

Generates λ as described in Model desctiptions/Main equations/Photoperiodic effect (page 763) in the origional paper.

Parameter: lat, the latitude

Returns: array of normalised change in day length

Main.SugarKelp.get_ind — Method

SugarKelp.get_ind(val, list, tol)

Function that finds the index in the list with the closest value to val. Error is thrown if no result is within tollerance, tol.

Parameters:

val: the value searching for

list: the list to search

tol: tollerance of search

Returns:

`index of closest valuea

Main.SugarKelp.interp_deps — Method

SugarKelp.interp_deps(arr, origional_depths, desired_depths, invalid_val)

Function to interpolate a 4D array in the last dimension. Useful for lineaising the depth steps of the arrays from Copurnicus as they have increasingly corse step sizes.

Parameters:

arr: the array to interpolate where the 3rd dimension is the one to interpolate along

origional_depths: the values of the 3rd dimensions coordinates (the origional depths of the data)

desired_depths: the desired coordinates/depths

invalid_val: the fill value to replace with NaN in the array (as we are searching it anyway)

Returns: new array with interpolated 3rd dimension

Main.SugarKelp.p — Method

SugarKelp.p(temp,irr)

Solves Equation 10, the gross photosynthesis.

Parameters:

temp: temperature /°C

irr: irradiance /mol photons/m²/day

Returns: gross photosynthesis function

Main.SugarKelp.r — Method

SugarKelp.r(temp,μ,j,resp_model)

Solves Equation 14 (2012) if respmodel=1, or Equation 2 (2013) if respmodel=2

Parameters:

temp: temperature /°C

μ: specific growth rate

j: specific nitrate uptake rate

resp_model: choice of resparation model (see description)

Returns: respiration function

Main.SugarKelp.solve — Function

SugarKelp.solve(t_i, nd, u, temp, irr, ex_n, lat, a_0, n_0, c_0, params="src/parameters/origional.jl", resp_model=1, dt=1, dataframe=true)

Solves the model for some set of parametetrs and returns the ODE library solution as well as a dataframe of the useful results.

Parameters:

t_i: th estart time (in days since the start of the interpolation objects "day zero")

nd: the number of days to run for

u: interpolation object (IO) of water speed

temp: IO of temperature

ex_n: IO of external nitrate concentration

lat: latitude, relivant for the change of day length

a_0: initial area

n_0: initial nitrogen reserve (gN/gSW)

c_0: initial carbon reserve (gC/gSW)

params: string of the path to a parameters file, defaults to the 2012 values. Also supplied is 2013 in src/parameters/2013.jl or you can copy and vary them

resp_model: choice of respiration model, 1 (default) uses the 2012 version and 2 uses the modifcations from the 2013 paper

dt: the time step size to use (see equations! note), default is 1 day (seems small enough)

dataframe: output as a dataframe, default to true. Alternative is an array (faster)

λ_arr: array of normalised change of day length, defaults to nothing which generates the default one

Returns:

solution: the ODE library solution

results: dataframe or array of area/nitrogen reserve/carbon reserve/total nitrate update. All others useful quantities can be easily derived.

Main.SugarKelp.solvegrid — Function

SugarKelp.solvegrid(t_i, nd, a_0, n_0, c_0, arr_lon, arr_lat, arr_dep, arr_time, no3, temp, u, par_data, kd_data, att = nothing, params="src/parameters/origional.jl", resp_model=1, dt=1, progress=true)

Solve the model for a (spacially) fixed grid of inputs.

Parameters:

t_i: start day (w.r.t. t=0 in time)

nd: number of days to run for

a_0: initial area/dm^2

n_0: initial nitrogen reserve/gN/gSW

c_0: initial carbon reserve/gC/gSW

arr_lon: longitudes for no3 and temp

arr_lat: latitudes for no3 and temp

arr_dep: depths for no3 and temp

arr_time: time for no3 and temp

no3: array of no3 concentration in lon,lat,depth,time

temp: array of temp concentration in lon,lat,depth,time

u: array of water speed in lon,lat,depth,time

par_data: array of:

`par values in lon,lat,time

`corresponding time

`par fill value

kd_data: array of:

kd values in lon,lat,time

corresponding time

kd fill value

att: array of light attenuation coefficients (PAR(z)=PAR(z=0)*att) in lon,lat,depth,time. Defaults to nothing and kd is used instead

params: string of the path to a parameters file, defaults to the 2012 values. Also supplied is 2013 in src/parameters/2013.jl or you can copy and vary them

resp_model: choice of respiration model, 1 (default) uses the 2012 version and 2 uses the modifcations from the 2013 paper

dt: the time step size to use (see equations! note), default is 1 day (seems small enough)

progress: option to update progress at each level (when multithreading not accurate but useful), defaults to true

Returns: results as an array of (area/nitrogen/carbon/nitrate update, lon, lat, depth, time)

Notes: par and kd need their own time coordinates and fill value because they come from satelite observation which are temporally sparse so need to be checked and interpolated in time for each point. On the other hand temp and no3 (that I'm using) are from Copurnicus' models so if a point has a value at some time it will at all times.

no3,temp and u need to be of the same shape and size and with the values corresponding to the same position/time.

Main.SugarKelp.solver! — Method

SugarKelp.solver!(y, params, t)

Interface for OrdinaryDiffEq library, extracts current enviromental variable values and solves equations for each timestep.

Parameters:

y: the current state of the system as a vector (area, nitrate reserve, carbon reserve).

params: the variable parameters of the model:

u_arr: interpolation object of the water speed in time

temp_arr: interpolation object of the temperature in time

irr_arr: interpolation object of the irradianec in time

ex_n_arr: interpolation object of the external nitrate concentration in time

λ_arr: array of the normalised change in day length

resp_model: the choice of respiration model, 1 is the origional from the 2012 paper and 2 is the modified version in 2013

dt: the time step length, this is important as it is used in the "extreme carbon limit" part of the equations, see NB.

t: the current time (with respect to the time in the interpolations)

Returns: array of da,dn,dc,j*a - the rate of a/n/c and the nitrate uptake rate

Note: These equations must be solved with an algorithm with fixed time steps and known, constant sub timestep lengths. This is because the "extreme carbon limit" element is only implientable with known and fixed timesteps as the value of a must be changed to a particular value rather than changing the deriviative. This can only be done (within the framework of the ODE library) by setting the derivitive to (X(next)-X(old))/dt

Main.SugarKelp.ν — Method

SugarKelp.ν(a)

Solves Equation 6, the specific frond erosion rate.

Parameters: a: area /dm^2

Returns: specific frond erosion rate